Fix removal of props
If an element has `[Html.Attributes.disabled True]` and then switches to
`[]`, the `disabled` property is supposed to be set to `false`. However,
this didn’t work because I had mixed up `props` and `prevProps`. I tried
to read `props["disabled"]` which is `undefined` (since `disabled` isn’t
set anymore), but should read `prevProps["disabled"]` (which is the
value `false`).